requirements.txt
gitpython
groq
fastapi
uvicorn[standard]